+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ApplicationForm/ApproveTermsSection/approveChildSupportTermsSubSection.ts
@@ -1,0 +1,14 @@
+import { buildCustomField, buildSubSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+
+export const approveChildSupportTermsSubSection = buildSubSection({
+  id: 'approveChildSupportTerms',
+  title: m.childSupport.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'approveChildSupportTerms',
+      title: m.childSupport.general.pageTitle,
+      component: 'ChildSupport',
+    }),
+  ],
+})
