+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ApplicationForm/ArrangementSection/index.ts
@@ -1,0 +1,17 @@
+import { buildSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+import { residenceChangeReasonSubSection } from './residenceChangeSubSection'
+import { confirmResidenceChangeInfoSubSection } from './confirmResidenceChangeInfoSubSection'
+import { childSupportPaymentsSubSection } from './childSupportPaymentsSubSection'
+import { durationSubSection } from './durationSubSection'
+
+export const arrangementSection = buildSection({
+  id: 'arrangement',
+  title: m.section.arrangement,
+  children: [
+    residenceChangeReasonSubSection,
+    confirmResidenceChangeInfoSubSection,
+    childSupportPaymentsSubSection,
+    durationSubSection,
+  ],
+})
