+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ApplicationForm/ArrangementSection/residenceChangeSubSection.ts
@@ -1,0 +1,14 @@
+import { buildCustomField, buildSubSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+
+export const residenceChangeReasonSubSection = buildSubSection({
+  id: 'residenceChangeReason',
+  title: m.reason.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'residenceChangeReason',
+      title: m.reason.general.pageTitle,
+      component: 'Reason',
+    }),
+  ],
+})
