+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ApplicationForm/BackgroundInformationSection/contactSubSection.ts
@@ -1,0 +1,16 @@
+import { buildCustomField, buildSubSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+import { contactInfoIds } from '../../../fields/ContactInfo'
+
+export const contactSubSection = buildSubSection({
+  id: 'contact',
+  title: m.contactInfo.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'contactInfo',
+      title: m.contactInfo.general.pageTitle,
+      childInputIds: contactInfoIds,
+      component: 'ContactInfo',
+    }),
+  ],
+})
