+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ApplicationForm/BackgroundInformationSection/index.ts
@@ -1,0 +1,10 @@
+import { buildSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+import { selectChildInCustodySubSection } from './selectChildInCustodySubSection'
+import { contactSubSection } from './contactSubSection'
+
+export const backgroundInformationSection = buildSection({
+  id: 'backgroundInformation',
+  title: m.section.backgroundInformation,
+  children: [selectChildInCustodySubSection, contactSubSection],
+})
