+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ConfirmationForms/ApplicationConfirmation.ts
@@ -5,7 +5,7 @@ import {
 } from '@island.is/application/core'
 import { Form } from '@island.is/application/types'
 import Logo from '@island.is/application/templates/family-matters-core/assets/Logo'
-import * as m from '../lib/messages'
+import * as m from '../../lib/messages'
 
 export const ApplicationConfirmation: Form = buildForm({
   id: 'ApplicationConfirmation',
