+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ParentBForm/contactSection/index.ts
@@ -1,0 +1,18 @@
+import { buildCustomField, buildSection } from '@island.is/application/core'
+import { contactInfoParentBIds } from '../../../fields/ContactInfoParentB'
+import { ApproveContract } from '../../../lib/dataSchema'
+import * as m from '../../../lib/messages'
+
+export const contactSection = buildSection({
+  condition: (answers) => answers.acceptContract === ApproveContract.Yes,
+  id: 'contact',
+  title: m.contactInfo.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'contactInfo',
+      title: m.contactInfo.general.pageTitle,
+      childInputIds: contactInfoParentBIds,
+      component: 'ContactInfoParentB',
+    }),
+  ],
+})
