+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ParentBForm/parentBIntroSection/index.ts
@@ -1,0 +1,14 @@
+import { buildCustomField, buildSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+
+export const parentBIntroSection = buildSection({
+  id: 'parentBIntro',
+  title: m.parentBIntro.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'acceptContract',
+      title: m.parentBIntro.general.pageTitle,
+      component: 'ParentBIntro',
+    }),
+  ],
+})
