+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ParentBForm/termsParentBSection/approveChildSupportTermsParentBSubSection.ts
@@ -1,0 +1,14 @@
+import { buildCustomField, buildSubSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+
+export const approveChildSupportTermsParentBSubSection = buildSubSection({
+  id: 'approveChildSupportTermsParentB',
+  title: m.childSupport.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'approveChildSupportTermsParentB',
+      title: m.childSupport.general.pageTitle,
+      component: 'ChildSupport',
+    }),
+  ],
+})
