+++ libs/application/templates/family-matters/children-residence-change-v2/src/forms/ParentBForm/termsParentBSection/approveTermsParentBSubSection.ts
@@ -1,0 +1,14 @@
+import { buildCustomField, buildSubSection } from '@island.is/application/core'
+import * as m from '../../../lib/messages'
+
+export const approveTermsParentBSubSection = buildSubSection({
+  id: 'approveTermsParentB',
+  title: m.terms.general.sectionTitle,
+  children: [
+    buildCustomField({
+      id: 'approveTermsParentB',
+      title: m.terms.general.pageTitle,
+      component: 'Terms',
+    }),
+  ],
+})
